projects/DMML#03/requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
streamlit
plotly

langchain
langchain-community
langchain-text-splitters

faiss-cpu

sentence-transformers

groq

python-dotenv
pypdf